chore: setup documetation and release process
