+++ README.md
@@ -9,6 +9,5 @@ Since the scope of this repo should stay very limited, tests are not included on
 > :warning: **Testing Terraform provider will create real ressources**: Be very mindful of the costs and to not disrupt critical infrastructure!
 
 ```bash
-cd `my_provider`
 TF_TEST_GCP_PROJECT_ID="GCP_PROJECT_TO_RUN_TEST_IN" TF_ACC=1 go test -v ./...
 ```
