+++ docs/data-sources/firestore_document.md
@@ -1,0 +1,27 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "passculture_firestore_document Data Source - passculture"
+subcategory: ""
+description: |-
+  
+---
+
+# passculture_firestore_document (Data Source)
+
+
+
+
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `collection` (String) Firestore collection
+- `database` (String) Firestore database
+- `document_id` (String) Document ID
+- `project` (String) Project ID
+
+### Read-Only
+
+- `fields` (Dynamic) Document fields
