+++ docs/index.md
@@ -1,0 +1,16 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "passculture Provider"
+subcategory: ""
+description: |-
+  
+---
+
+# passculture Provider
+
+
+
+
+
+<!-- schema generated by tfplugindocs -->
+## Schema
