--- terraform-provider-firestore/internal/provider/provider.go
@@ -1,44 +1,0 @@
-package provider
-
-import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/ephemeral"
-	"github.com/hashicorp/terraform-plugin-framework/provider"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-)
-
-var _ provider.Provider = (*firestoreProvider)(nil)
-
-type firestoreProvider struct{}
-
-func New() func() provider.Provider {
-	return func() provider.Provider {
-		return &firestoreProvider{}
-	}
-}
-
-func (p *firestoreProvider) Configure(ctx context.Context, req provider.ConfigureRequest, resp *provider.ConfigureResponse) {
-}
-
-func (p *firestoreProvider) Metadata(ctx context.Context, req provider.MetadataRequest, resp *provider.MetadataResponse) {
-	resp.TypeName = "firestore"
-}
-
-func (p *firestoreProvider) Resources(ctx context.Context) []func() resource.Resource {
-	return []func() resource.Resource{}
-}
-
-func (p *firestoreProvider) EphemeralResources(ctx context.Context) []func() ephemeral.EphemeralResource {
-	return []func() ephemeral.EphemeralResource{}
-}
-
-func (p *firestoreProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
-	return []func() datasource.DataSource{
-		NewFirestoreDocumentDataSource,
-	}
-}
-
-func (p *firestoreProvider) Schema(ctx context.Context, req provider.SchemaRequest, resp *provider.SchemaResponse) {
-}
