+++ tools.go
@@ -1,0 +1,9 @@
+//go:build generate
+
+package tools
+
+import (
+	_ "github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs"
+)
+
+//go:generate go run github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs generate --provider-dir . -provider-name passculture
